test: organize viewport regression modules
